Add experimental support to channel and load information inside AIM
This commit is still work-in-progress.
This new version of AIM includes support to an experimental
enhancement to CAM messages, to include an additional optional
container with information on the channel established between
the nodes and on their computational resources (CPU, RAM).
This support may be dropped in the future to enable again
the reception of standard CAMv2.
This update is still highly experimental and will come to a
more stable point with the next few commits.
Please note that, currently, these enhanced CAMs can only be
generated leveraging the OCABS project.