Initial decoding and logging support for DENMs
This commit introduces two new options (-D and -P) which enable
decoding and logging for DENM messages, in addition to CAM messages.
As opposed to CAMs, DENMs are used for the time being just for
logging (if -D is specified), and not for updating the internal
database (which will come as a future update).
The -P option will make AIM log the current position, speed and
heading for each received DENM, if a GNSS receiver is available
through gpsd. Thus, to be available, AIM needs to be compiled
with libgps support, through two new Makefile targets:
compilePC_gpsd if you compile on the device which will run AIM
compileAPU_gpsd if you need to cross-compile for OpenWrt